Fix: upload gdrive delete
